Fix numpy RuntimeWarning 'invalid value encountered in cast' in factors.py
Root cause: ma_structure_raw values (0–100 integers) created an int64
numpy array. When safe_percentile_rank() called np.full_like(arr, np.nan),
numpy attempted to cast NaN into the int64 array, triggering:

  RuntimeWarning: invalid value encountered in cast
    multiarray.copyto(res, fill_value, casting='unsafe')

Fix: Specify dtype=float when creating the ma_structures array, matching
the convention of all other factor arrays (momentum_prices, etc.) which
are naturally float from their computations.

+++ scripts/factors.py
@@ -328,7 +328,7 @@ def run(weights=None, as_of_date=None):
     momentum_prices = np.array([r['momentum_price_raw'] for r in results])
     momentum_vols = np.array([r['momentum_vol_raw'] for r in results])
     trend_adxs = np.array([r['trend_adx_raw'] for r in results])
-    ma_structures = np.array([r['ma_structure_raw'] for r in results])
+    ma_structures = np.array([r['ma_structure_raw'] for r in results], dtype=float)
     pullbacks = np.array([r['pullback_raw'] for r in results])
     rsis = np.array([r['rsi_raw'] for r in results])
     liquiditys = np.array([r['liquidity_raw'] for r in results])
